Add files via upload
All you need this main file and then update the Firebase configuration part. When you hit the main page, you're in a lobby. Hit the create room button on the top right, then the game will start, and you can give the room code to a buddy for them to join you :)